incident_env/server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openenv-core[core]>=0.2.2
